Revert the KeyedListLike, just use KeyedList
