+++ scalding-core/src/main/scala/com/twitter/scalding/typed/CoGrouped2.scala
@@ -23,10 +23,12 @@ import com.twitter.scalding._
 
 import scala.collection.JavaConverters._
 
-class CoGrouped2[K,V,W,R](left: Grouped[K,V],
+class CoGrouped2[+K,V,W,+R](left: Grouped[K,V],
   right: Grouped[K,W],
   joiner: (K, Iterator[V], Iterable[W]) => Iterator[R])
   extends KeyedList[K,R] with java.io.Serializable {
+
+  type This[+K, +R] = CoGrouped2[K,V,W,R]
 
   override lazy val toTypedPipe : TypedPipe[(K,R)] = {
     // Actually make a new coGrouping:
@@ -72,7 +74,7 @@ class CoGrouped2[K,V,W,R](left: Grouped[K,V],
     TypedPipe.from[(K,R)](pipeWithRed, ('key, 'value))
   }
 
-  override def mapValueStream[U](fn: Iterator[R] => Iterator[U]): KeyedList[K,U] = {
+  override def mapValueStream[U](fn: Iterator[R] => Iterator[U]) = {
     new CoGrouped2[K,V,W,U](left, right, {(k,vit,wit) => fn(joiner(k,vit,wit))})
   }
 }
